broke post_init tests into multiple test cases
